Enabled ShorthandShouldBeThrownBySpec.scala under Scala 3 build, changed related matchers code to use Scala 3 macro for position.
